+++ CHANGELOG.md
@@ -1,8 +1,16 @@
 # CHANGELOG
 
+## v0.1.2
+
+- Added support for Firefox
+- Added option to scan all media on a page
+- Various updates
+
 ## v0.1.1
 
 - IFrames, including nested IFrames and IFrames within shadow-roots, are now scanned for media elements
+- Display trust list logo in popup (if available)
+- Various updates
 
 ## v0.1.0